Fixing the PersistentBag issue, comparing with equals().
When I create a new instance of Autor class, with its List<Livro>
attribute left as null, if I Persist() it and Find() it back, that
attribute that was null is retrieved as a Non Null PersistentBag.

With this behavior, if I use equals() in this two instances, they are
representing the same logical object, but they give different Equals()
and Hash() responses.

What I did here was to test if that attribute InstanceOf PersistentBag
and its Empty at the same time, and I make it Null as the original
object were, making Equals() work as I was expecting